Finished addinf the rest of the new view files
