+++ cs_app/views/__init__.py
@@ -1,3 +1,6 @@
 from .admin_views import *
 from .authorization_views import *
 from .generate_report_views import *
+from .directions_views import *
+from .change_database_views import *
+from .account_information_views import *
