+++ cs_app/views/account_information_views.py
@@ -1,0 +1,64 @@
+from django.http import JsonResponse
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import render
+
+
+@login_required
+def account_information_view(request):
+    user = request.user
+    context = {
+        "user": user,
+    }
+
+    return render(request, "account_information.html", context)
+
+
+@login_required
+def update_name_view(request):
+    if request.method == "POST":
+        first_name = request.POST.get("first_name")
+        last_name = request.POST.get("last_name")
+
+        user = request.user
+
+        # Process the data as needed
+        user.first_name = first_name
+        user.last_name = last_name
+        user.save()
+
+        # Return a JSON response
+        return JsonResponse({"message": "Data received successfully"})
+
+    # Return an error response if the request method is not POST
+    return JsonResponse({"error": "Invalid request method"}, status=400)
+
+
+@login_required
+def update_email_view(request):
+    if request.method == "POST":
+        email = request.POST.get("email")
+
+        user = request.user
+
+        user.email = email
+        user.username = email
+        user.save()
+
+        return JsonResponse({"message": "Data received successfully"})
+
+    return JsonResponse({"error": "Invalid request method"}, status=400)
+
+
+@login_required
+def update_password_view(request):
+    if request.method == "POST":
+        password = request.POST.get("password")
+
+        user = request.user
+
+        user.set_password(password)
+        user.save()
+
+        return JsonResponse({"message": "Data received successfully"})
+
+    return JsonResponse({"error": "Invalid request method"}, status=400)
